Enhanced instruction and added BEQL

+++ src/cpu/cp0/cp0.rs
@@ -12,7 +12,7 @@ impl CP0 {
         self.reg_config.power_on_reset();
     }
 
-    pub fn write_reg(&mut self, index: u32, data: u64) {
+    pub fn write_reg(&mut self, index: u8, data: u64) {
         match index {
             12 => {
                 self.reg_status = (data as u32).into();
